[TMail] JAMES-4210 Move TMail PLAIN SASL support to a shared SASL maven module
Extract TMail PLAIN SASL mechanism and factory from imap-extensions into sasl-extensions so IMAP and future SMTP adoption can reuse the same delegation-aware implementation.
