+++ tmail-backend/sasl-extensions/pom.xml
@@ -28,54 +28,31 @@
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>imap-extensions</artifactId>
-    <name>Twake Mail :: IMAP :: Extensions</name>
+    <artifactId>sasl-extensions</artifactId>
+    <name>Twake Mail :: SASL :: Extensions</name>
 
-   <dependencies>
-       <dependency>
-           <groupId>${project.groupId}</groupId>
-           <artifactId>sasl-extensions</artifactId>
-       </dependency>
-       <dependency>
-           <groupId>${project.groupId}</groupId>
-           <artifactId>team-mailboxes</artifactId>
-       </dependency>
-       <dependency>
-           <groupId>${james.groupId}</groupId>
-           <artifactId>james-server-guice-imap</artifactId>
-       </dependency>
-       <dependency>
-           <groupId>org.apache.james.protocols</groupId>
-           <artifactId>protocols-imap</artifactId>
-       </dependency>
-       <dependency>
-           <groupId>org.apache.james.protocols</groupId>
-           <artifactId>protocols-sasl</artifactId>
-       </dependency>
-       <dependency>
-           <groupId>org.apache.james.protocols</groupId>
-           <artifactId>protocols-imap</artifactId>
-           <type>test-jar</type>
-           <scope>test</scope>
-       </dependency>
-       <dependency>
-           <groupId>com.google.inject</groupId>
-           <artifactId>guice</artifactId>
-       </dependency>
-       <dependency>
-           <groupId>commons-io</groupId>
-           <artifactId>commons-io</artifactId>
-       </dependency>
-       <dependency>
-           <groupId>${james.groupId}</groupId>
-           <artifactId>testing-base</artifactId>
-           <scope>test</scope>
-       </dependency>
-       <dependency>
-           <groupId>${james.groupId}</groupId>
-           <artifactId>apache-james-mailbox-api</artifactId>
-           <type>test-jar</type>
-           <scope>test</scope>
-       </dependency>
-   </dependencies>
+    <dependencies>
+        <dependency>
+            <groupId>org.apache.james.protocols</groupId>
+            <artifactId>protocols-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.james.protocols</groupId>
+            <artifactId>protocols-sasl</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-lang3</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>${james.groupId}</groupId>
+            <artifactId>testing-base</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>jcl-over-slf4j</artifactId>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
 </project>
